Implemented a Redis repository. Also refactored the existing repository interface and base class.
