+++ src/Repositories/LocalRepository.php
@@ -4,360 +4,26 @@ namespace Caffeinated\Modules\Repositories;
 
 class LocalRepository extends Repository
 {
+
     /**
-     * Get all modules.
+     * Returns the entire module repository content as a collection.
      *
      * @return Collection
      */
-    public function all()
+    public function load()
     {
-        return $this->getCache()->sortBy('order');
+        return $this->getCache();
     }
 
     /**
-     * Get all module slugs.
+     * Saves the content to the repository.
      *
-     * @return Collection
+     * @param $content
+     * @return int|bool a non-zero or true value on success, or false on failure.
      */
-    public function slugs()
-    {
-        $slugs = collect();
-
-        $this->all()->each(function ($item, $key) use ($slugs) {
-            $slugs->push(strtolower($item['slug']));
-        });
-
-        return $slugs;
-    }
-
-    /**
-     * Get modules based on where clause.
-     *
-     * @param string $key
-     * @param mixed  $value
-     *
-     * @return Collection
-     */
-    public function where($key, $value)
-    {
-        return collect($this->all()->where($key, $value)->first());
-    }
-
-    /**
-     * Sort modules by given key in ascending order.
-     *
-     * @param string $key
-     *
-     * @return Collection
-     */
-    public function sortBy($key)
-    {
-        $collection = $this->all();
-
-        return $collection->sortBy($key);
-    }
-
-    /**
-     * Sort modules by given key in ascending order.
-     *
-     * @param string $key
-     *
-     * @return Collection
-     */
-    public function sortByDesc($key)
-    {
-        $collection = $this->all();
-
-        return $collection->sortByDesc($key);
-    }
-
-    /**
-     * Determines if the given module exists.
-     *
-     * @param string $slug
-     *
-     * @return bool
-     */
-    public function exists($slug)
-    {
-        return $this->slugs()->contains($slug);
-    }
-
-    /**
-     * Returns count of all modules.
-     *
-     * @return int
-     */
-    public function count()
-    {
-        return $this->all()->count();
-    }
-
-    /**
-     * Get a module property value.
-     *
-     * @param string $property
-     * @param mixed  $default
-     *
-     * @return mixed
-     */
-    public function get($property, $default = null)
-    {
-        list($slug, $key) = explode('::', $property);
-
-        $module = $this->where('slug', $slug);
-
-        return $module->get($key, $default);
-    }
-
-    /**
-     * Set the given module property value.
-     *
-     * @param string $property
-     * @param mixed  $value
-     *
-     * @return bool
-     */
-    public function set($property, $value)
-    {
-        list($slug, $key) = explode('::', $property);
-
-        $cachePath = $this->getCachePath();
-        $cache = $this->getCache();
-        $module = $this->where('slug', $slug);
-
-        if (isset($module[$key])) {
-            unset($module[$key]);
-        }
-
-        $module[$key] = $value;
-
-        $module = collect([$module['basename'] => $module]);
-
-        $merged = $cache->merge($module);
-        $content = json_encode($merged->all(), JSON_PRETTY_PRINT);
-
-        return $this->files->put($cachePath, $content);
-    }
-
-    /*
-    |--------------------------------------------------------------------------
-    | Initialization Methods
-    |--------------------------------------------------------------------------
-    |
-    */
-
-    /**
-     * Get all initialized modules.
-     *
-     * @return Collection
-     */
-    public function initialized()
-    {
-        return $this->all()->where('initialized', true);
-    }
-
-    /**
-     * Get all uninitialized modules.
-     *
-     * @return Collection
-     */
-    public function uninitialized()
-    {
-        return $this->all()->where('initialized', false);
-    }
-
-    /**
-     * Check if specified module is initialized.
-     *
-     * @param string $slug
-     *
-     * @return bool
-     */
-    public function isInitialized($slug)
-    {
-        $module = $this->where('slug', $slug);
-
-        return $module['initialized'] === true;
-    }
-
-    /**
-     * Check if specified module is uninitialized.
-     *
-     * @param string $slug
-     *
-     * @return bool
-     */
-    public function isUninitialized($slug)
-    {
-        $module = $this->where('slug', $slug);
-
-        return $module['initialized'] === false;
-    }
-
-    /**
-     * Initialize the specified module.
-     *
-     * @param string $slug
-     *
-     * @return bool
-     */
-    public function initialize($slug)
-    {
-        if (parent::initialize($slug)) {
-            return $this->set($slug.'::initialized', true);
-        } else {
-            return false;
-        }
-    }
-
-    /**
-     * Uninitialize the specified module.
-     *
-     * @param string $slug
-     *
-     * @return bool
-     */
-    public function uninitialize($slug)
-    {
-        if (parent::uninitialize($slug)) {
-            return $this->set($slug.'::initialized', false);
-        } else {
-            return false;
-        }
-    }
-
-    /*
-    |--------------------------------------------------------------------------
-    | Enabling Methods
-    |--------------------------------------------------------------------------
-    |
-    */
-
-    /**
-     * Get all enabled modules.
-     *
-     * @return Collection
-     */
-    public function enabled()
-    {
-        return $this->all()->where('enabled', true);
-    }
-
-    /**
-     * Get all disabled modules.
-     *
-     * @return Collection
-     */
-    public function disabled()
-    {
-        return $this->all()->where('enabled', false);
-    }
-
-    /**
-     * Check if specified module is enabled.
-     *
-     * @param string $slug
-     *
-     * @return bool
-     */
-    public function isEnabled($slug)
-    {
-        $module = $this->where('slug', $slug);
-
-        return $module['enabled'] === true;
-    }
-
-    /**
-     * Check if specified module is disabled.
-     *
-     * @param string $slug
-     *
-     * @return bool
-     */
-    public function isDisabled($slug)
-    {
-        $module = $this->where('slug', $slug);
-
-        return $module['enabled'] === false;
-    }
-
-    /**
-     * Enables the specified module.
-     *
-     * @param string $slug
-     *
-     * @return bool
-     */
-    public function enable($slug)
-    {
-        if (parent::enable($slug)) {
-            return $this->set($slug.'::enabled', true);
-        } else {
-            return false;
-        }
-    }
-
-    /**
-     * Disables the specified module.
-     *
-     * @param string $slug
-     *
-     * @return bool
-     */
-    public function disable($slug)
-    {
-        if (parent::disable($slug)) {
-            return $this->set($slug.'::enabled', false);
-        } else {
-            return false;
-        }
-    }
-
-    /*
-    |--------------------------------------------------------------------------
-    | Optimization Methods
-    |--------------------------------------------------------------------------
-    |
-    */
-
-    /**
-     * Update cached repository of module information.
-     *
-     * @return bool
-     */
-    public function optimize()
+    public function save($content)
     {
         $cachePath = $this->getCachePath();
-
-        $cache = $this->getCache();
-        $basenames = $this->getAllBasenames();
-        $modules = collect();
-
-        $basenames->each(function ($module, $key) use ($modules, $cache) {
-            $basename = collect(['basename' => $module]);
-            $temp = $basename->merge(collect($cache->get($module)));
-            $manifest = $temp->merge(collect($this->getManifest($module)));
-
-            $modules->put($module, $manifest);
-        });
-
-        $modules->each(function ($module) {
-            $module->put('id', crc32($module->get('slug')));
-
-            if (!$module->has('enabled')) {
-                $module->put('enabled', config('modules.enabled', true));
-            }
-
-            if (!$module->has('order')) {
-                $module->put('order', 9001);
-            }
-
-            return $module;
-        });
-
-        $content = json_encode($modules->all(), JSON_PRETTY_PRINT);
-
         return $this->files->put($cachePath, $content);
     }
