+++ src/Repositories/Repository.php
@@ -37,12 +37,154 @@ abstract class Repository implements RepositoryContract
         $this->files = $files;
     }
 
+    /*
+    |--------------------------------------------------------------------------
+    | Optimization Methods
+    |--------------------------------------------------------------------------
+    |
+    */
+
+    /**
+     * Update cached repository of module information.
+     *
+     * @return bool
+     */
+    public function optimize()
+    {
+
+        $cache = $this->load();
+        $basenames = $this->getAllBasenames();
+        $modules = collect();
+
+        $basenames->each(function ($module, $key) use ($modules, $cache) {
+            $basename = collect(['basename' => $module]);
+            $temp = $basename->merge(collect($cache->get($module)));
+            $manifest = $temp->merge(collect($this->getManifest($module)));
+
+            $modules->put($module, $manifest);
+        });
+
+        $modules->each(function ($module) {
+            $module->put('id', crc32($module->get('slug')));
+
+            if (!$module->has('enabled')) {
+                $module->put('enabled', config('modules.enabled', true));
+            }
+
+            if (!$module->has('order')) {
+                $module->put('order', 9001);
+            }
+
+            return $module;
+        });
+
+        $content = json_encode($modules->all(), JSON_PRETTY_PRINT);
+
+        return $this->save($content);
+    }
+
+    /*
+    |--------------------------------------------------------------------------
+    | Collection Methods
+    |--------------------------------------------------------------------------
+    |
+    */
+
+    /**
+     * Get all modules.
+     *
+     * @return Collection
+     */
+    public function all()
+    {
+        return $this->load()->sortBy('order');
+    }
+
+    /**
+     * Get all module slugs.
+     *
+     * @return Collection
+     */
+    public function slugs()
+    {
+        $slugs = collect();
+
+        $this->all()->each(function ($item, $key) use ($slugs) {
+            $slugs->push(strtolower($item['slug']));
+        });
+
+        return $slugs;
+    }
+
+    /**
+     * Get modules based on where clause.
+     *
+     * @param string $key
+     * @param mixed  $value
+     *
+     * @return Collection
+     */
+    public function where($key, $value)
+    {
+        return collect($this->all()->where($key, $value)->first());
+    }
+
+    /**
+     * Sort modules by given key in ascending order.
+     *
+     * @param string $key
+     *
+     * @return Collection
+     */
+    public function sortBy($key)
+    {
+        $collection = $this->all();
+
+        return $collection->sortBy($key);
+    }
+
+    /**
+     * Sort modules by given key in ascending order.
+     *
+     * @param string $key
+     *
+     * @return Collection
+     */
+    public function sortByDesc($key)
+    {
+        $collection = $this->all();
+
+        return $collection->sortByDesc($key);
+    }
+
+    /**
+     * Determines if the given module exists.
+     *
+     * @param string $slug
+     *
+     * @return bool
+     */
+    public function exists($slug)
+    {
+        return $this->slugs()->contains($slug);
+    }
+
+    /**
+     * Returns count of all modules.
+     *
+     * @return int
+     */
+    public function count()
+    {
+        return $this->all()->count();
+    }
+
     /**
      * Get all module basenames.
      *
      * @return array
      */
-    protected function getAllBasenames()
+    public function getAllBasenames()
     {
         $path = $this->getPath();
 
@@ -58,6 +200,13 @@ abstract class Repository implements RepositoryContract
             return collect([]);
         }
     }
+
+    /*
+    |--------------------------------------------------------------------------
+    | Manifests and their paths Methods
+    |--------------------------------------------------------------------------
+    |
+    */
 
     /**
      * Get a module's manifest contents.
@@ -88,7 +237,7 @@ abstract class Repository implements RepositoryContract
      *
      * @return string
      */
-    public function getPath()
+    protected function getPath()
     {
         return $this->path ?: $this->config->get('modules.path');
     }
@@ -100,7 +249,7 @@ abstract class Repository implements RepositoryContract
      *
      * @return object $this
      */
-    public function setPath($path)
+    protected function setPath($path)
     {
         $this->path = $path;
 
@@ -114,7 +263,7 @@ abstract class Repository implements RepositoryContract
      *
      * @return string
      */
-    public function getModulePath($slug)
+    protected function getModulePath($slug)
     {
         $module = studly_case(str_slug($slug));
 
@@ -147,8 +296,117 @@ abstract class Repository implements RepositoryContract
         return rtrim($this->config->get('modules.namespace'), '/\\');
     }
 
+    /*
+    |--------------------------------------------------------------------------
+    | Repository read and write Methods
+    |--------------------------------------------------------------------------
+    |
+    */
+
     /**
-     * Calls the initialize maintenance method for the specified module.
+     * Get a module property value.
+     *
+     * @param string $property
+     * @param mixed  $default
+     *
+     * @return mixed
+     */
+    public function get($property, $default = null)
+    {
+        list($slug, $key) = explode('::', $property);
+
+        $module = $this->where('slug', $slug);
+
+        return $module->get($key, $default);
+    }
+
+    /**
+     * Set the given module property value.
+     *
+     * @param string $property
+     * @param mixed  $value
+     *
+     * @return bool
+     */
+    public function set($property, $value)
+    {
+        list($slug, $key) = explode('::', $property);
+
+        $cache = $this->load();
+        $module = $this->where('slug', $slug);
+
+        if (isset($module[$key])) {
+            unset($module[$key]);
+        }
+
+        $module[$key] = $value;
+
+        $module = collect([$module['basename'] => $module]);
+
+        $merged = $cache->merge($module);
+        $content = json_encode($merged->all(), JSON_PRETTY_PRINT);
+
+        return $this->save($content);
+    }
+
+
+    /*
+    |--------------------------------------------------------------------------
+    | Initialization Methods
+    |--------------------------------------------------------------------------
+    |
+    */
+
+    /**
+     * Get all initialized modules.
+     *
+     * @return Collection
+     */
+    public function initialized()
+    {
+        return $this->all()->where('initialized', true);
+    }
+
+    /**
+     * Get all uninitialized modules.
+     *
+     * @return Collection
+     */
+    public function uninitialized()
+    {
+        return $this->all()->where('initialized', false);
+    }
+
+    /**
+     * Check if specified module is initialized.
+     *
+     * @param string $slug
+     *
+     * @return bool
+     */
+    public function isInitialized($slug)
+    {
+        $module = $this->where('slug', $slug);
+
+        return $module['initialized'] === true;
+    }
+
+    /**
+     * Check if specified module is uninitialized.
+     *
+     * @param string $slug
+     *
+     * @return bool
+     */
+    public function isUninitialized($slug)
+    {
+        $module = $this->where('slug', $slug);
+
+        return $module['initialized'] === false;
+    }
+
+    /**
+     * Initialize the specified module.
      *
      * @param string $slug
      *
@@ -156,11 +414,15 @@ abstract class Repository implements RepositoryContract
      */
     public function initialize($slug)
     {
-        return $this->callMaintenanceMethod($slug, 'initialize');
+        if ($this->callMaintenanceMethod($slug, 'initialize')) {
+            return $this->set($slug.'::initialized', true);
+        } else {
+            return false;
+        }
     }
 
     /**
-     * Calls the uninitialize maintenance method for the specified module.
+     * Uninitialize the specified module.
      *
      * @param string $slug
      *
@@ -168,11 +430,70 @@ abstract class Repository implements RepositoryContract
      */
     public function uninitialize($slug)
     {
-        return $this->callMaintenanceMethod($slug, 'uninitialize');
+        if ($this->callMaintenanceMethod($slug, 'uninitialize')) {
+            return $this->set($slug.'::initialized', false);
+        } else {
+            return false;
+        }
+    }
+
+    /*
+    |--------------------------------------------------------------------------
+    | Enabling Methods
+    |--------------------------------------------------------------------------
+    |
+    */
+
+    /**
+     * Get all enabled modules.
+     *
+     * @return Collection
+     */
+    public function enabled()
+    {
+        return $this->all()->where('enabled', true);
     }
 
     /**
-     * Calls the enable maintenance method for the specified module.
+     * Get all disabled modules.
+     *
+     * @return Collection
+     */
+    public function disabled()
+    {
+        return $this->all()->where('enabled', false);
+    }
+
+    /**
+     * Check if specified module is enabled.
+     *
+     * @param string $slug
+     *
+     * @return bool
+     */
+    public function isEnabled($slug)
+    {
+        $module = $this->where('slug', $slug);
+
+        return $module['enabled'] === true;
+    }
+
+    /**
+     * Check if specified module is disabled.
+     *
+     * @param string $slug
+     *
+     * @return bool
+     */
+    public function isDisabled($slug)
+    {
+        $module = $this->where('slug', $slug);
+
+        return $module['enabled'] === false;
+    }
+
+    /**
+     * Enables the specified module.
      *
      * @param string $slug
      *
@@ -180,11 +501,15 @@ abstract class Repository implements RepositoryContract
      */
     public function enable($slug)
     {
-        return $this->callMaintenanceMethod($slug, 'enable');
+        if ($this->callMaintenanceMethod($slug, 'enable')) {
+            return $this->set($slug.'::enabled', true);
+        } else {
+            return false;
+        }
     }
 
     /**
-     * Calls the disable maintenance method for the specified module.
+     * Disables the specified module.
      *
      * @param string $slug
      *
@@ -192,7 +517,11 @@ abstract class Repository implements RepositoryContract
      */
     public function disable($slug)
     {
-        return $this->callMaintenanceMethod($slug, 'disable');
+        if ($this->callMaintenanceMethod($slug, 'disable')) {
+            return $this->set($slug.'::enabled', false);
+        } else {
+            return false;
+        }
     }
 
     /**
